# docs/swagger.yaml
basePath: /api/v1
definitions:
  database.Todo:
    properties:
      completed:
        type: boolean
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      userId:
        type: string
    type: object
  database.TodoCreate:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  database.TodoPatch:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        minLength: 1
        type: string
    type: object
  database.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  main.loginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  main.loginResponse:
    properties:
      token:
        type: string
    type: object
  main.registerRequest:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
info:
  contact: {}
  description: A simple Go web API for managing todos.
  title: Go Web API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.loginResponse'
      summary: Logs in a user
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.User'
      summary: Registers a new user
      tags:
      - users
  /api/v1/todos:
    get:
      consumes:
      - application/json
      description: Get all todos for user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Todo'
            type: array
      security:
      - BearerAuth: []
      summary: Get all todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo
      parameters:
      - description: Todo object
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/database.TodoCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Todo'
      security:
      - BearerAuth: []
      summary: Create a new todo
      tags:
      - todos
  /api/v1/todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/database.Todo'
      security:
      - BearerAuth: []
      summary: Delete a todo
      tags:
      - todos
    patch:
      consumes:
      - application/json
      description: Update a todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Todo object
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/database.TodoPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Todo'
      security:
      - BearerAuth: []
      summary: Update a todo
      tags:
      - todos
produces:
- application/json
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: Enter your bearer token in the format "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"